backend/requirements.txt
# Core dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0

# JWT Authentication
PyJWT==2.8.0
Flask-JWT-Extended==4.6.0

# Password hashing
Werkzeug==3.0.1
bcrypt==4.1.2

# Redis for rate limiting (optional)
redis==5.0.1

# Date and time handling
python-dateutil==2.8.2

# Validation and serialization
marshmallow==3.20.2
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Testing
pytest==7.4.4
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development tools
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Production server
gunicorn==21.2.0

# Environment and configuration
click==8.1.7
requests==2.31.0
orjson==3.9.10